requirements.txt
fastapi[all]
sqlalchemy
psycopg2-binary
pydantic
uvicorn 
passlib[bcrypt]
python-jose
